+++ CHANGELOG.md
@@ -6,6 +6,10 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/).
 
 ## [Unreleased]
 
+### Fixed
+
+- Fixed document symbol crash on incomplete functions
+
 ## [1.16.1] - 2023-01-30
 
 ### Fixed